appstore/requirements.txt
codegen==1.0
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
Markdown==3.1
patterns==0.3
Pillow==6.0.0
psycopg2-binary==2.7.7
PyJWT==1.7.1
pytz==2018.9
